renamed connector to mcpselectionslice
