+++ libs/designer/src/lib/ui/mcp/panel/connector/SelectionPanel.tsx
@@ -8,7 +8,7 @@ import { useMcpPanelStyles } from '../styles';
 import { useIntl } from 'react-intl';
 import { bundleIcon, Dismiss24Filled, Dismiss24Regular } from '@fluentui/react-icons';
 import { useCallback } from 'react';
-import { clearAllSelections } from '../../../../core/state/mcp/connector/connectorSlice';
+import { clearAllSelections } from '../../../../core/state/mcp/mcpselectionslice';
 
 const CloseIcon = bundleIcon(Dismiss24Filled, Dismiss24Regular);
